Project以外のデータ入力

+++ database/migrations/2023_08_02_062942_create_ship_owners_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('ship_id')->constrained();
             $table->string('owner_name');
-            $table->integer('ratio');
+            $table->float('ratio');
             $table->timestamps();
 
         });
